set integer field with SetIntField not SetLongField, bug caught by wonderful other-endianness of the G4
darcs-hash:20061109044200-0c629-a0a7024b5ddd51ea8394bc11c8ad22ccd9c75b5e.gz

+++ src/org/sqlite/DB.c
@@ -162,13 +162,13 @@ static xCall(
 
     (*env)->SetLongField(env, func, fld_context, fromref(context));
     (*env)->SetLongField(env, func, fld_value, value ? fromref(value) : 0);
-    (*env)->SetLongField(env, func, fld_args, args);
+    (*env)->SetIntField(env, func, fld_args, args);
 
     (*env)->CallVoidMethod(env, func, method);
 
     (*env)->SetLongField(env, func, fld_context, 0);
     (*env)->SetLongField(env, func, fld_value, 0);
-    (*env)->SetLongField(env, func, fld_args, 0);
+    (*env)->SetIntField(env, func, fld_args, 0);
 
     // check if xFunc threw an Exception
     if ((*env)->ExceptionCheck(env)) xFunc_error(context, env);
